move code into single file
